refactor(namespace): NamespaceStore abstract base, RAM and Redis in separate files
