+++ typescript/packages/core/src/workspace/mount/namespace.test.ts
@@ -16,7 +16,7 @@ import { describe, expect, it } from 'vitest'
 import { RAMResource } from '../../resource/ram/ram.ts'
 import { CycleError } from '../../utils/path.ts'
 import { Workspace } from '../workspace.ts'
-import { RAMNamespaceStore } from './namespace_store.ts'
+import { RAMNamespaceStore } from './namespace_ram_store.ts'
 
 describe('Namespace facade (addressing)', () => {
   it('resolve delegates to the workspace resolver', async () => {
